ZipArchive - try to delete unnecessary file only if it exists

+++ src/main/java/VASSAL/tools/io/ZipArchive.java
@@ -453,7 +453,9 @@ public class ZipArchive implements FileArchive {
     // Replace old archive with temp archive.
     if (!tmpFile.renameTo(archiveFile)) {
       try {
-        FileUtils.forceDelete(archiveFile);
+        if (archiveFile.exists()) {
+          FileUtils.forceDelete(archiveFile);
+        }
         FileUtils.moveFile(tmpFile, archiveFile);
       }
       catch (IOException e) {
